Refactored getUptimeGraph into ActiveTimeGraph react component. Also updated Shadow and Ele graphs to use rolling average instead of cumulative average
