+++ src/analysis/retail/priest/shadow/modules/features/AlwaysBeCasting.tsx
@@ -3,29 +3,10 @@ import { formatPercentage } from 'common/format';
 import { ThresholdStyle, When } from 'parser/core/ParseResults';
 import CoreAlwaysBeCasting from 'parser/shared/modules/AlwaysBeCasting';
 import { STATISTIC_ORDER } from 'parser/ui/StatisticBox';
-import { GlobalCooldownEvent } from 'parser/core/Events';
-import getUptimeGraph, { UptimeHistoryEntry } from 'parser/shared/modules/getUptimeGraph';
 import { QualitativePerformance } from 'parser/ui/QualitativePerformance';
 
 class AlwaysBeCasting extends CoreAlwaysBeCasting {
   position = STATISTIC_ORDER.CORE(6);
-
-  uptimeHistory: UptimeHistoryEntry[] = [];
-
-  onGCD(event: GlobalCooldownEvent) {
-    const super_result = super.onGCD(event);
-
-    this.uptimeHistory.push({
-      timestamp: this.owner.currentTimestamp,
-      uptimePct: this.activeTimePercentage,
-    });
-
-    return super_result;
-  }
-
-  get graphSubsection() {
-    return getUptimeGraph(this.uptimeHistory, this.owner.fight.start_time);
-  }
 
   get suggestionThresholds() {
     return {
